app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
pillow==10.1.0
